more tests and removed keyactivity
