--- app/src/main/res/layout/activity_key.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="nl.tudelft.cs4160.trustchain_android.KeyActivity">
-
-    <android.widget.ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="40dp"
-                android:text="@string/private_key" />
-
-            <TextView
-                android:id="@+id/private_key"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
-
-            <Button
-                android:id="@+id/new_key"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/create_key" />
-
-            <Button
-                android:id="@+id/sign_data"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/sign_data" />
-
-            <Button
-                android:id="@+id/verify_sig"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/verify_sig" />
-
-            <TextView
-                android:id="@+id/signed_data"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
-        </LinearLayout>
-
-    </android.widget.ScrollView>
-
-
-</android.support.design.widget.CoordinatorLayout>
