--- app/src/test/java/nl/tudelft/cs4160/trustchain_android/databaseTest/TrustChainDBHelperTest.java
@@ -1,8 +1,0 @@
-package nl.tudelft.cs4160.trustchain_android.databaseTest;
-
-/**
- * Created by Boning on 12/17/2017.
- */
-
-public class TrustChainDBHelperTest {
-}
